Move common registry code for buildscripts registry lib
Prepare for CMK-14575
